+++ CHANGELOG.md
@@ -5,6 +5,7 @@
     * attendance\_limit\_includes\_leaders Boolean
   * Add limited\_attendance? method to Event
   * Add setting of a section's notepad
+  * Add updating of Activity
 
 ## Version 0.1.12